requirements.txt
Flask>=2.0
Pillow>=9.0
mutagen>=1.45
cryptography>=40.0
click>=8.0
colorama>=0.4
python-magic>=0.4.24
Flask-Limiter>=2.0
gunicorn>=20.0
requests>=2.0
exifread>=2.0
